HDDS-11993. Move OzoneAdmin to ozone-tools
