+++ hadoop-ozone/dev-support/intellij/runConfigurations/ScmRoles.xml
@@ -16,7 +16,7 @@
 -->
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="ScmRoles" type="Application" factoryName="Application">
-    <option name="MAIN_CLASS_NAME" value="org.apache.hadoop.hdds.cli.OzoneAdmin" />
+    <option name="MAIN_CLASS_NAME" value="org.apache.hadoop.ozone.admin.OzoneAdmin" />
     <module name="ozone-tools" />
     <option name="PROGRAM_PARAMETERS" value="-conf=hadoop-ozone/dev-support/intellij/ozone-site-ha.xml scm roles -id=scm-group" />
     <option name="VM_PARAMETERS" value="-Dlog4j.configuration=file:hadoop-ozone/dev-support/intellij/log4j.properties" />
